// src/api/generated/models/post-search-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 4차 프로젝트 7팀 API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface PostSearchRequest
 */
export interface PostSearchRequest {
    /**
     * 
     * @type {string}
     * @memberof PostSearchRequest
     */
    'category'?: string;
    /**
     * 
     * @type {number}
     * @memberof PostSearchRequest
     */
    'minPrice'?: number;
    /**
     * 
     * @type {number}
     * @memberof PostSearchRequest
     */
    'maxPrice'?: number;
    /**
     * 
     * @type {boolean}
     * @memberof PostSearchRequest
     */
    'saleStatus'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof PostSearchRequest
     */
    'keyword'?: string;
    /**
     * 
     * @type {string}
     * @memberof PostSearchRequest
     */
    'place'?: string;
}